# docs/requirements/api-spec.yaml
openapi: 3.1.0
info:
  title: Staff Procurement Portal API
  version: 0.1.0
  summary: Internal vendor procurement portal for staff
  description: |
    Read-only API backing the Staff Procurement Portal demo. Derived directly
    from `docs/requirements/REQUIREMENTS.md` v0.1 and `docs/requirements/FUNCTIONAL_DESIGN.md`
    v0.1.

    Identity is established by ID.me OAuth 2.0 (authorisation code flow). The backend
    performs the code exchange, validates the ID token, looks the user up in the Oracle
    `STAFF` table, and issues a JWT in an `HttpOnly; SameSite=Lax` cookie named
    `session`. All `/api/vendors/*` endpoints require this cookie. All access to
    Bank details and audit logging are out of scope per ADR-012 — no `BANK_DETAILS` field is returned, and no `AUDIT_LOG` table exists.

    Naming, layer order, and conventions follow `AI_CONTEXT.md`. RBAC follows the
    role/level matrix in `REQUIREMENTS.md` §2.2.
  contact:
    name: Cloud & Things (C&T) — Staff Procurement Portal team
  license:
    name: Proprietary
  x-source-documents:
    - docs/requirements/REQUIREMENTS.md
    - docs/requirements/FUNCTIONAL_DESIGN.md
  x-status: DRAFT
servers:
  - url: http://{ec2_public_ip}
    description: Demo on AWS EC2 (HTTP for the demo; HTTPS is a phase-2 NFR)
    variables:
      ec2_public_ip:
        default: localhost
        description: EC2 public IP supplied at deploy time
  - url: http://localhost
    description: Local Docker Compose development
...
security:
  - cookieAuth: []

# ─────────────────────────────────────────────────────────────────────────────
# Paths
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /api/vendors:
    get:
      tags: [vendors]
      summary: List vendors (RBAC-filtered)
      description: |
        Returns the role-appropriate slice of `PROCUREMENT_ITEMS`. Per
        [ADR-015](../adr/ADR-015-role-model-simplification.md), the level
        dimension was removed; role is the single authority axis.

        | Role                      | Statuses returned | Columns returned |
        | ------------------------- | ----------------- | ---------------- |
        | `REGULAR_STAFF`           | APPROVED only     | `item_id, vendor_name, item_name, category, contact_name` (no `contact_email` per ADR-013) |
        | `PROCUREMENT_SUPERVISOR`  | All statuses      | + `status, unit_price` (no `bank_details` per ADR-012; no `contact_email` per ADR-013) |

        > Note: today every authenticated session receives the `admin`
        > variant — per-role narrowing is the follow-up to ADR-015. The
        > schema variants (`staff_l1` / `staff_l2` / `admin`) remain in
        > place so the SPA's `VendorTable` can pick columns from
        > `row.variant` when narrowing lands.

        An empty list is a valid 200 response — the SPA renders the empty state.
      operationId: listVendors
      responses:
        "200":
          description: Vendor list, shaped per the caller's role
          content:
            application/json:
              schema:
                type: array
                items:
                  oneOf:
                    - $ref: "#/components/schemas/VendorListItemL1"
                    - $ref: "#/components/schemas/VendorListItemL2"
                    - $ref: "#/components/schemas/VendorListItemAdmin"
                  discriminator:
                    propertyName: variant
                    mapping:
                      staff_l1: "#/components/schemas/VendorListItemL1"
                      staff_l2: "#/components/schemas/VendorListItemL2"
                      admin:    "#/components/schemas/VendorListItemAdmin"
              examples:
                staff_l1_empty:
                  summary: LEVEL 1 staff, no approved vendors yet
                  value: []
                staff_l2:
                  summary: LEVEL 2 staff
                  value:
                    - variant: staff_l2
                      item_id: 12
                      vendor_name: "Dell Technologies"
                      item_name: "Laptop fleet (Latitude 5440)"
                      category: "Technology"
                      contact_name: "Pat Lee"
                admin:
                  summary: Admin (LEVEL 3)
                  value:
                    - variant: admin
                      item_id: 12
                      vendor_name: "Dell Technologies"
                      item_name: "Laptop fleet (Latitude 5440)"
                      category: "Technology"
                      contact_name: "Pat Lee"
                      status: "APPROVED"
                      unit_price: 1129.00
        "401":
          $ref: "#/components/responses/Unauthenticated"
  /api/vendors/{vendor_id}:
    get:
      tags: [vendors]
      summary: Get a single vendor's full record (admin only)
      description: |
        Admin-only. Returns the full `PROCUREMENT_ITEMS` row (excluding
        `BANK_DETAILS` per ADR-012) plus the new `approved_at` date when
        `status == "APPROVED"`. *Previous drafts wrote an `AUDIT_LOG` row on
        every response containing `bank_details`; the entire audit-log
        requirement was removed by ADR-012.*
      operationId: getVendorDetail
      parameters:
        - in: path
          name: vendor_id
          required: true
          schema: { type: integer, minimum: 1 }
          description: "`PROCUREMENT_ITEMS.ITEM_ID`"
      responses:
        "200":
          description: Vendor detail (admin only)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VendorDetail" }
              example:
                variant: admin_detail
                item_id: 12
                vendor_name: "Dell Technologies"
                item_name: "Laptop fleet (Latitude 5440)"
                category: "Technology"
                contact_name: "Pat Lee"
                contact_email: "pat.lee@dell.example.com"
                status: "APPROVED"
                unit_price: 1129.00
                approved_at: "2026-03-14"
                created_date: "2026-02-28T14:03:22Z"
                updated_date: "2026-03-14T09:18:05Z"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          description: Authenticated but not authorised (caller is not `PROCUREMENT_SUPERVISOR`)
          headers:
            X-Auth-Reason:
              required: true
              schema: { type: string, enum: [ROLE_FORBIDDEN] }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example:
                detail: "You do not have permission to view this resource."
        "404":
          description: No vendor exists with the given `vendor_id`
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example:
                detail: "Vendor not found."
        "500":
          description: |
            Server-side failure (Oracle query error, connection drop, etc.).
            The browser-facing body never contains stack traces. *(The
            "audit write failure" case from earlier drafts was removed by
            ADR-012 along with the rest of the audit-log requirement.)*
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example:
                detail: "Something went wrong. Please try again."

# ─────────────────────────────────────────────────────────────────────────────
# Components
# ─────────────────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: |
        JWT (HS256, claims `sub=staff_id, role, iat, exp, iss, aud`)
        in an `HttpOnly; SameSite=Lax` cookie. TTL configurable via `JWT_TTL_HOURS`
        (default 4). Issued by `POST /api/auth/callback`, cleared by
        `POST /api/auth/logout`. The `session` cookie is the **only** accepted
        credential — no bearer header.
  responses:
    Unauthenticated:
      description: |
        No session cookie, or the cookie is invalid/expired. The SPA's API client
        treats any 401 as a global event and redirects to `/?reason=session_expired`.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            detail: "Session invalid or expired."
...
  schemas:
...
    VendorStatus:
      type: string
      enum: [PENDING, UNDER_REVIEW, APPROVED, REJECTED]
    VendorListItemL1:
      type: object
      description: |
        Narrowest list-item shape. Reserved for future per-role narrowing
        (see [ADR-015](../adr/ADR-015-role-model-simplification.md));
        today every authenticated session receives the `admin` variant.
      required: [variant, item_id, vendor_name, item_name, category]
      additionalProperties: false
      properties:
        variant:
          type: string
          const: staff_l1
        item_id:
          type: integer
          description: "`PROCUREMENT_ITEMS.ITEM_ID`"
        vendor_name:
          type: string
        item_name:
          type: string
        category:
          type: string
    VendorListItemL2:
      type: object
      description: |
        Intermediate list-item shape (carries `contact_name`). Reserved
        for future per-role narrowing (see [ADR-015](../adr/ADR-015-role-model-simplification.md));
        today every authenticated session receives the `admin` variant.
        `contact_email` was removed from this schema by ADR-013 — email lives only in `VendorDetail`.
      required: [variant, item_id, vendor_name, item_name, category, contact_name]
      additionalProperties: false
      properties:
        variant:
          type: string
          const: staff_l2
        item_id: { type: integer }
        vendor_name: { type: string }
        item_name: { type: string }
        category: { type: string }
        contact_name:
          type: string
          nullable: true
    VendorListItemAdmin:
      type: object
      description: |
        List item visible to `PROCUREMENT_SUPERVISOR` (per ADR-015 —
        previously `ADMIN`).
        `bank_details` was removed by ADR-012; `contact_email` was removed by ADR-013 (lives only in `VendorDetail`).
      required:
        - variant
        - item_id
        - vendor_name
        - item_name
        - category
        - contact_name
        - status
        - unit_price
      additionalProperties: false
      properties:
        variant:
          type: string
          const: admin
        item_id: { type: integer }
        vendor_name: { type: string }
        item_name: { type: string }
        category: { type: string }
        contact_name:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/VendorStatus"
        unit_price:
          type: number
          format: double
          nullable: true
          minimum: 0
    VendorDetail:
      type: object
      description: |
        Full vendor record for `PROCUREMENT_SUPERVISOR` (per ADR-015). The **only** schema in the API surface
        that returns `contact_email` (per [ADR-013](../adr/ADR-013-api-responses-match-ui-display.md))
        — list endpoints return contact name only. `bank_details` was removed
        from this schema by ADR-012, along with the audit-log row this endpoint
        used to write.
      required:
        - variant
        - item_id
        - vendor_name
        - item_name
        - category
        - contact_name
        - contact_email
        - status
        - unit_price
        - approved_at
        - created_date
        - updated_date
      additionalProperties: false
      properties:
        variant:
          type: string
          const: admin_detail
        item_id: { type: integer }
        vendor_name: { type: string }
        item_name: { type: string }
        category: { type: string }
        contact_name:
          type: string
          nullable: true
        contact_email:
          type: string
          format: email
          nullable: true
        status:
          $ref: "#/components/schemas/VendorStatus"
        unit_price:
          type: number
          format: double
          nullable: true
        approved_at:
          type: string
          format: date
          nullable: true
          description: |
            Date the vendor's status last transitioned to `APPROVED`. `null` for
            vendors that have never been approved. Requires the
            `PROCUREMENT_ITEMS.APPROVED_AT` column added by ADR-006 (see
            FUNCTIONAL_DESIGN §2 conflict C-06).
        created_date:
          type: string
          format: date-time
        updated_date:
          type: string
          format: date-time

    # ─── Errors ────────────────────────────────────────────────────────────
    Error:
      type: object
      description: |
        FastAPI default error envelope. Browser-facing copy only — stack traces and
        internal codes never leak (NFR-12). Sub-reason for 403s is delivered via the
        `X-Auth-Reason` response header, not in the body.
      required: [detail]
      additionalProperties: false
      properties:
        detail:
          type: string
          description: User-facing error message. Safe to display verbatim.